Matchmaking page restructured
