+++ src/assets/styles/Common.styles.ts
@@ -2,6 +2,21 @@ import React from "react";
 import styled, { ThemeProps } from "styled-components";
 import Theme from "../../theme";
 
+const CrossButton = styled.div`
+    position: absolute;
+    left: 20px;
+    top: 20px;
+`;
+
+const WhiteBar = styled.div`
+    width: 80%;
+    color: white;
+    height: 1px;
+    background: white;
+    margin: 30px auto 10px;
+    border-radius: 15px;
+`;
+
 const StyledBody = styled.div`
     max-width: 500px;
     margin: 56px auto;
@@ -14,4 +29,4 @@ const OdourlessWrapper = styled((props) => React.createElement(props.component, 
     color: ${({ theme }: ThemeProps<typeof Theme>) => theme.palette.common.white};
 `;
 
-export { StyledBody, OdourlessWrapper };
+export { StyledBody, OdourlessWrapper, CrossButton, WhiteBar };
