+++ src/components/logo/Logo.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from "react";
 import { LogoDiv, LogoImg } from "./Logo.styles";
 
-const Logo = (props: Logo): ReactElement => {
+const Logo = (props: LogoProps): ReactElement => {
     return (
         <LogoDiv style={props.styles}>
             <LogoImg src={props.imgUrl} alt="logo" />
@@ -11,7 +11,7 @@ const Logo = (props: Logo): ReactElement => {
 
 export default Logo;
 
-interface Logo {
+interface LogoProps {
     imgUrl: string;
     styles: any;
 }
